Split the interactive-flow provider manager into generic + auth-consumer (#294)
Extract the shared third-party-provider machinery (authorize, callback dispatch,
token exchange, claim resolution, route-by-purpose, re-authentication confirm) out
of the OAuth2-authorize consumer into a generic
business.manager.flow.InteractiveFlowSessionOAuth2ProviderManager, and move the
sign-up/merge/invitation logic behind a ProviderUserEstablisher seam implemented by
InteractiveAuthFlowSessionProviderEstablisher.

The manager now serves OAUTH2_AUTHORIZE, REAUTHENTICATION and (upcoming) LINK_PROVIDER
and is driven by the generic ProvidersController, so its flow.auth package and Auth
prefix no longer fit. Behaviour-preserving; the test is split accordingly.
